extend mother-to-child linkage post submission handler to allow editing
