--- packages/esm-ohri-pmtct-app/src/utils/ptracker-forms-helpers.ts
@@ -1,11 +1,0 @@
-export const generateInfantPTrackerId = (
-    fieldId: string,
-    motherPtrackerId: string
-  ): string | undefined => {
-    if (!fieldId || !motherPtrackerId) return;
-    return fieldId === "infantPtrackerid"
-      ? motherPtrackerId + "1"
-      : fieldId.includes("_")
-      ? motherPtrackerId.concat(fieldId.split("_")[1])
-      : undefined;
-  };
